Issue #14937: Migrated section 4.8.3.2 No C Style Array Declarations to chapter wise testing
