+++ src/xdocs/google_style.xml
@@ -1228,7 +1228,7 @@
                   </a>
                 </td>
                 <td>
-                  <a href="styleguides/google-java-style-20180523/javaguide.html#s4.8.3.2-array-declarations">
+                  <a href="styleguides/google-java-style-20220203/javaguide.html#s4.8.3.2-array-declarations">
                     4.8.3.2 No C-style array declarations</a>
                 </td>
                 <td>
@@ -1241,7 +1241,7 @@
                   <a href="https://github.com/search?q=path%3Asrc%2Fmain%2Fresources%20path%3A**%2Fgoogle_checks.xml+repo%3Acheckstyle%2Fcheckstyle+ArrayTypeStyle">
                     config</a>
                   <br />
-                  <a href="https://github.com/checkstyle/checkstyle/blob/master/src/it/java/com/google/checkstyle/test/chapter4formatting/rule4832nocstylearray/ArrayTypeStyleTest.java">
+                  <a href="https://github.com/checkstyle/checkstyle/blob/master/src/it/java/com/google/checkstyle/test/chapter4formatting/rule4832nocstylearray/NoCstyleArrayDeclarationsTest.java">
                     test</a>
                 </td>
               </tr>
